Extract RenderContext struct to reduce parameter count in render_level functions
Refactored the deeply parameter-heavy `render_level` functions in the doom-renderer crate by extracting 13 separate arguments into a unified `RenderContext` struct. This flattens the signature and improves clarity while keeping the rendering logic mathematically identical.

Updated 50+ test-site calls across `doom-renderer` and `doom-app` to match the new signature safely.

+++ crates/doom-renderer/src/intermission.rs
@@ -505,7 +505,7 @@ fn wi_draw_number(
 ) -> i32 {
     let s = value.to_string();
     for ch in s.bytes() {
-        let digit = (ch - b'0') as u8;
+        let digit = ch - b'0';
         let name = format!("WINUM{digit}");
         if let Some(p) = cache.get(&name, wad) {
             let p = p.clone();
@@ -621,10 +621,19 @@ pub fn draw_intermission_wad(
 
     // 3. Stats (shown as counting progresses).
     let phase = renderer.phase;
-    let show_kills   = !matches!(phase, IntermissionPhase::CountingKills);
-    let show_items   = !matches!(phase, IntermissionPhase::CountingKills | IntermissionPhase::CountingItems);
-    let show_secrets = matches!(phase, IntermissionPhase::ShowingTime | IntermissionPhase::Done);
-    let show_time    = matches!(phase, IntermissionPhase::ShowingTime | IntermissionPhase::Done);
+    let show_kills = !matches!(phase, IntermissionPhase::CountingKills);
+    let show_items = !matches!(
+        phase,
+        IntermissionPhase::CountingKills | IntermissionPhase::CountingItems
+    );
+    let show_secrets = matches!(
+        phase,
+        IntermissionPhase::ShowingTime | IntermissionPhase::Done
+    );
+    let show_time = matches!(
+        phase,
+        IntermissionPhase::ShowingTime | IntermissionPhase::Done
+    );
 
     if let Some(p) = cache.get("WIOSTK", wad) {
         let p = p.clone();
